Add files via upload
In this project we take the MacBook Air was the first product in Mac series launched with the M1 chip produced by Apple itself, which replaced the Intel Processors in Apple Products. This gives a relevance in analyzing the customer reviews about the new product from a e-commerce website.For getting the data from websites we done the web scrapping. Finding number of duplicated reviews deleting the duplicated reviews. Converting to lower case, removing stopwords, removing punctuation, tokenization.At the end we deploy the model by using Streamlit.